+++ R/compute_variances.R
@@ -16,7 +16,10 @@
 
   if (faminfo$family %in% c("truncated_nbinom1")) {
     if (verbose) {
-      warning(format_message(sprintf("Truncated negative binomial families are currently not supported by `%s`.", name_fun)), call. = FALSE)
+      warning(format_message(sprintf(
+        "Truncated negative binomial families are currently not supported by `%s`.", 
+        name_fun)
+      ), call. = FALSE)
     }
     return(NA)
   }
@@ -223,7 +226,7 @@
     }
 
     vcorr <- compact_list(list(vc1, vc2))
-    names(vcorr) <- c("cond", "zi")[1:length(vcorr)]
+    names(vcorr) <- c("cond", "zi")[seq_along(vcorr)]
 
     vals <- list(
       beta = lme4::fixef(x),
@@ -250,14 +253,14 @@
       vc = vcorr,
       re = list(lme4::ranef(x))
     )
-    names(vals$re) <- re_names[1:length(vals$re)]
+    names(vals$re) <- re_names[seq_along(vals$re)]
 
     # nlme
     # ---------------------------
   } else if (inherits(x, "lme")) {
     re_names <- find_random(x, split_nested = TRUE, flatten = TRUE)
     comp_x <- get_modelmatrix(x)
-    rownames(comp_x) <- 1:nrow(comp_x)
+    rownames(comp_x) <- seq_len(nrow(comp_x))
     if (.is_nested_lme(x)) {
       vals_vc <- .get_nested_lme_varcorr(x)
       vals_re <- lme4::ranef(x)
@@ -301,7 +304,7 @@
     # ---------------------------
   } else if (inherits(x, "brmsfit")) {
     comp_x <- get_modelmatrix(x)
-    rownames(comp_x) <- 1:nrow(comp_x)
+    rownames(comp_x) <- seq_len(nrow(comp_x))
     vc <- lapply(names(lme4::VarCorr(x)), function(i) {
       element <- lme4::VarCorr(x)[[i]]
       if (i != "residual__") {
@@ -383,7 +386,9 @@
 # is supported or not
 .badlink <- function(link, family, verbose = TRUE) {
   if (verbose) {
-    warning(format_message(sprintf("Model link '%s' is not yet supported for the %s distribution.", link, family)), call. = FALSE)
+    warning(format_message(sprintf(
+      "Model link '%s' is not yet supported for the %s distribution.", link, family)
+    ), call. = FALSE)
   }
   return(NA)
 }
@@ -875,8 +880,8 @@
   # make sure we have identical subcomponents between random and
   # fixed effects
   fe <- compact_list(fe[c("conditional", "zero_inflated")])
-  if (length(rs) > length(fe)) rs <- rs[1:length(fe)]
-  if (length(fe) > length(rs)) fe <- fe[1:length(rs)]
+  if (length(rs) > length(fe)) rs <- rs[seq_along(fe)]
+  if (length(fe) > length(rs)) fe <- fe[seq_along(rs)]
 
   all(mapply(function(r, f) all(r %in% f), rs, fe, SIMPLIFY = TRUE))
 }
@@ -932,7 +937,7 @@
       if (length(missig_rnd_slope)) {
         # sanity check
         to_remove <- c()
-        for (j in 1:length(out)) {
+        for (j in seq_along(out)) {
           # identical random slopes might have different names, so
           # we here check if random slopes from correlated and uncorrelated
           # are duplicated (i.e. their difference is 0 - including a tolerance)