+++ R/get_df.R
@@ -177,6 +177,19 @@ get_df.cgam <- function(x, type = "residual", ...) {
   if (type == "model") {
     .model_df(x)
   } else {
+    # x$resid_df_obs
+    # new in cgam 1.18
+    stats::df.residual(x)
+  }
+}
+
+
+#' @export
+get_df.cgamm <- function(x, type = "residual", ...) {
+  type <- match.arg(tolower(type), choices = c("residual", "model"))
+  if (type == "model") {
+    .model_df(x)
+  } else {
     x$resid_df_obs
   }
 }
@@ -453,7 +466,7 @@ get_df.systemfit <- function(x, type = "residual", ...) {
   f <- find_formula(x)
   system_names <- names(f)
 
-  for (i in 1:length(system_names)) {
+  for (i in seq_along(system_names)) {
     dfs <- rep(s[[i]]$df[2], length(params[[i]]))
     df_names <- rep(names(params[i]), length(params[[i]]))
     df <- c(df, stats::setNames(dfs, df_names))