+++ R/get_predicted.R
@@ -127,8 +127,13 @@
 #' }
 #'
 #' \subsection{Bayesian and Bootstrapped models and iterations}{
-#' For predictions based on multiple iterations, for instance in the case of Bayesian models and bootstrapped predictions, the function used to compute the centrality (point-estimate predictions) can be modified via the `centrality_function` argument. For instance, `get_predicted(model, centrality_function = stats::median)`. The default is `mean`.
-#' Individual draws can be accessed by running `iter <- as.data.frame(get_predicted(model))`, and their iterations can be reshaped into a long format by `bayestestR::reshape_iterations(iter)`.
+#' For predictions based on multiple iterations, for instance in the case of Bayesian
+#' models and bootstrapped predictions, the function used to compute the centrality
+#' (point-estimate predictions) can be modified via the `centrality_function`
+#' argument. For instance, `get_predicted(model, centrality_function = stats::median)`.
+#' The default is `mean`. Individual draws can be accessed by running
+#' `iter <- as.data.frame(get_predicted(model))`, and their iterations can be
+#' reshaped into a long format by `bayestestR::reshape_iterations(iter)`.
 #' }
 #'
 #' @seealso [get_datagrid()]
@@ -212,7 +217,10 @@ get_predicted.default <- function(x,
 
   # may fail due to invalid "dot_args", so try shorter argument list
   if (is.null(predictions)) {
-    predictions <- tryCatch(do.call("predict", compact_list(list(x, newdata = args$data, type = args$type))), error = function(e) NULL)
+    predictions <- tryCatch(
+      do.call("predict", compact_list(list(x, newdata = args$data, type = args$type))),
+      error = function(e) NULL
+    )
   }
 
   # still fails? try fitted()
@@ -333,6 +341,15 @@ get_predicted.glm <- get_predicted.lm
 get_predicted.lrm <- get_predicted.default
 
 
+
+# MASS: rlm -----------------------------------------------------
+# =======================================================================
+
+# these objects inherit from `lm`, but `get_predicted.lm` do not return
+# confidence intervals.
+
+#' @export
+get_predicted.rlm <- get_predicted.default
 
 
 # survival: survreg -----------------------------------------------------
@@ -549,7 +566,13 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
       ci_data[!se_col] <- lapply(ci_data[!se_col], .get_predict_transform_response, response = response)
       predictions <- .get_predict_transform_response(predictions, response = response)
       if ("iterations" %in% names(attributes(predictions))) {
-        attr(predictions, "iterations") <- as.data.frame(sapply(attributes(predictions)$iterations, .get_predict_transform_response, response = response))
+        attr(predictions, "iterations") <- as.data.frame(
+          sapply(
+            attributes(predictions)$iterations,
+            .get_predict_transform_response,
+            response = response
+          )
+        )
       }
     }
   }
@@ -576,7 +599,7 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
   if (is.matrix(predictions) && ncol(predictions) > 1) {
     predictions <- as.data.frame(predictions)
     vary <- colnames(predictions)
-    predictions$Row <- 1:nrow(predictions)
+    predictions$Row <- seq_len(nrow(predictions))
     # if we have any focal predictors, add those as well, so we have
     # the associated levels/values for "Row"
     if (!is.null(args$data)) {
@@ -641,7 +664,7 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
 
   # Format draws
   draws <- as.data.frame(t(draws$t))
-  names(draws) <- paste0("iter_", 1:ncol(draws))
+  names(draws) <- paste0("iter_", seq_len(ncol(draws)))
 
   .get_predicted_centrality_from_draws(x, draws, ...)
 }
@@ -663,7 +686,7 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
     iter_stacked <- apply(iter, 1, c)
     predictions <- data.frame(
       # rows repeated for each response level
-      Row = rep(1:ncol(iter), times = dim(iter)[3]),
+      Row = rep(seq_len(ncol(iter)), times = dim(iter)[3]),
       # response levels repeated for each row
       Response = rep(dimnames(iter)[[3]], each = dim(iter)[2]),
       Predicted = apply(iter_stacked, 1, centrality_function),
@@ -679,7 +702,7 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
     predictions <- apply(iter, 1, centrality_function)
   }
   # Rename iterations
-  names(iter) <- paste0("iter_", 1:ncol(iter))
+  names(iter) <- paste0("iter_", seq_len(ncol(iter)))
   # Store as attribute
   attr(predictions, "iterations") <- iter
   predictions