+++ R/get_varcov.R
@@ -32,8 +32,10 @@
 #' @param vcov_args List of arguments to be passed to the function identified by
 #'   the `vcov` argument. This function is typically supplied by the **sandwich**
 #'   or **clubSandwich** packages. Please refer to their documentation (e.g.,
-#'   `?sandwich::vcovHAC`) to see the list of available arguments.
-
+#'   `?sandwich::vcovHAC`) to see the list of available arguments. If no estimation
+#'   type (argument `type`) is given, the default type for `"HC"` (or `"vcovHC"`)
+#'   equals the default from the **sandwich** package; for type `"CR"` (or
+#'   `"vcoCR"`), the default is set to `"CR3"`.
 #' @param verbose Toggle warnings.
 #' @param ... Currently not used.
 #'
@@ -801,7 +803,7 @@ get_varcov.lqmm <- function(x, ...) {
   np <- length(find_parameters(x, flatten = TRUE))
 
   if (length(dim(sc$Cov)) == 3) {
-    vc <- lapply(1:length(x$tau), function(i) {
+    vc <- lapply(seq_along(x$tau), function(i) {
       .x <- sc$Cov[, , i][1:np, 1:np]
       if (.is_negativ_matrix(.x, ...)) {
         .x <- .fix_negative_matrix(.x)