+++ R/get_varcov_sandwich.R
@@ -61,6 +61,13 @@
       "residual", "wild", "mammen", "webb"
     ))) {
       vcov_args[["type"]] <- vcov_fun
+    } else if (is.null(vcov_fun)) {
+      # set defaults
+      vcov_args[["type"]] <- switch(
+        vcov_fun,
+        "CR" = "CR3",
+        NULL
+      )
     }
   }
 
@@ -125,7 +132,7 @@
 
   # extract variance-covariance matrix
   if (!inherits(.vcov, "matrix")) {
-    msg <- sprintf("Unable to extract a variance-covariance matrix for model object of class `%s`. Different values of the `vcov` argument trigger calls to the `sandwich` or `clubSandwich` packages in order to extract the matrix (see `?insight::get_varcov`). Your model or the requested estimation type may not be supported by one or both of those packages.", class(x)[1])
+    msg <- sprintf("Unable to extract a variance-covariance matrix for model object of class `%s`. Different values of the `vcov` argument trigger calls to the `sandwich` or `clubSandwich` packages in order to extract the matrix (see `?insight::get_varcov`). Your model or the requested estimation type may not be supported by one or both of those packages, or you were missing one or more required arguments in `vcov_args` (like `cluster`).", class(x)[1])
     stop(format_message(msg), call. = FALSE)
   }