+++ tests/testthat/test-get_predicted.R
@@ -178,6 +178,18 @@ test_that("robust vcov", {
 })
 
 
+
+test_that("MASS::rlm", {
+  requiet("MASS")
+  mod <- rlm(mpg ~ hp + am, data = mtcars)
+  p <- get_predicted.default(mod)
+  expect_s3_class(p, "get_predicted")
+  p <- data.frame(p)
+  expect_true("CI_low" %in% colnames(p))
+})
+
+
+
 # Mixed --------------------------------------------------------------
 # =========================================================================